refactor(scripts): extract query tokenizer to shared module (issue 4301 prep)
cluster-orphan-queries.mjs's IT/EN/DE/FR tokenizer, stemmer, locale
detector, and role/region canonicalizer move to scripts/lib/query-tokenizer.mjs
so the upcoming PostHog internal-search mining script can reuse the exact
same logic instead of duplicating STOPWORDS/STEM_RULES/REGION_TOKENS
(AGENTS.md non-negotiable #6). Verified byte-identical output on a real run
against the committed clusters file (only diff was fresh source-data drift,
reverted).
