+++ scripts/cluster-orphan-queries.mjs
@@ -35,6 +35,15 @@
 import fs from 'node:fs';
 import path from 'node:path';
 import { fileURLToPath } from 'node:url';
+import {
+  normalize,
+  tokenize,
+  detectLocale,
+  splitRoleRegion,
+  canonicalizeRegion,
+  canonicalizeRole,
+  slugify,
+} from './lib/query-tokenizer.mjs';
 
 const __dirname = path.dirname(fileURLToPath(import.meta.url));
 const ROOT = path.resolve(__dirname, '..');
@@ -42,160 +51,11 @@ const DRY_RUN = process.argv.includes('--dry-run');
 
 const MIN_CLUSTER_IMPRESSIONS = 5;
 
-// ─── Stopword sets (IT/EN/DE/FR) ──────────────────────────────────
-const STOPWORDS = new Set([
-  // Italian
-  'a','al','alla','allo','alle','ai','agli','di','da','del','dello','della','dei','degli','delle',
-  'in','un','uno','una','con','per','il','lo','la','gli','le','e','ed','o','od','che','chi','ciò',
-  'ciao','come','cosa','cui','ne','non','più','meno','essere','io','mi','tu','ti','si','ci','vi',
-  'sono','loro','nostro','vostro','su','sul','sulla','sulle','sui','sugli','tra','fra','dopo',
-  'prima','senza','questa','questo','questi','queste','quel','quella','quello',
-  // English
-  'the','of','in','for','to','and','or','a','an','with','at','by','from','on','off','up','down',
-  'as','is','are','be','been','being','has','have','had','do','does','did','will','would','should',
-  'can','could','may','might','must','shall','not','no',
-  // German
-  'der','die','das','den','dem','des','ein','eine','einen','einem','einer','eines','und','oder',
-  'mit','bei','in','im','am','an','auf','zu','von','aus','für','durch','um','nach','vor','ohne',
-  'gegen','zwischen','über','unter','hinter','neben','sein','ist','bin','bist','sind','seid',
-  // French
-  'le','la','les','l','un','une','des','du','de','d','et','ou','où','pour','par','avec','sans',
-  'dans','en','sur','sous','vers','chez','aux','au','à','ce','ces','cette','ses','son','sa',
-  // Generic job words — kept OUT of stopwords because they're part of role signal
-]);
-
-// Minimal multi-language stemming: remove common plural/gender endings.
-const STEM_RULES = [
-  /innen?$/,   // DE fem. plural → root  (Mitarbeiterinnen → Mitarbeiter)
-  /euse$/,     // FR fem. suffix → eur   (chauffeuse → chauff)
-  /teur$/, /teuse$/,
-  /eurs?$/,    // FR → eur
-  /ieres?$/,   // IT/FR fem.
-  /iere$/,
-  /ori$/, /ore$/, /ori$/, /ice$/, /ici$/, // IT
-  /i$/, /e$/, /o$/, /a$/, // IT/generic final-vowel trim as very-last step
-  /ing$/,      // EN gerund
-  /s$/,        // EN plural
-];
-
-function stemToken(tok) {
-  if (!tok) return tok;
-  if (tok.length <= 3) return tok;
-  // Apply the first rule that strips at least one char and leaves ≥3 chars remaining.
-  for (const rule of STEM_RULES) {
-    const next = tok.replace(rule, '');
-    if (next.length >= 3 && next.length < tok.length) return next;
-  }
-  return tok;
-}
-
-// ─── Locale detection heuristics ──────────────────────────────────
-const LOCALE_HINTS = {
-  it: ['lavoro','lavori','offerta','offerte','assunzione','assume','assunzioni','aziende','stipendio','posti','posto','cerca','cerco','ticino','svizzera','italiani','lugano','mendrisio','chiasso','bellinzona','locarno','frontaliere','concorso','concorsi'],
-  de: ['jobs','job','stellen','stelle','arbeit','arbeits','schweiz','tessin','stellenangebote','stellenangebot','mitarbeiter','mitarbeiterin','suche','offene','stellensuche','als','bei','für'],
-  en: ['jobs','job','work','switzerland','swiss','ticino','employment','careers','career','vacancies','vacancy','hiring','engineer','developer','nurse','nurses'],
-  fr: ['emploi','emplois','travail','suisse','tessin','offres','offre','recherche','poste','postes','carriere','postuler','chauffeur'],
-};
-
-function detectLocale(tokens) {
-  const scores = { it: 0, de: 0, en: 0, fr: 0 };
-  for (const t of tokens) {
-    for (const [loc, hints] of Object.entries(LOCALE_HINTS)) {
-      if (hints.includes(t)) scores[loc] += 1;
-    }
-  }
-  let best = 'it';
-  let bestScore = -1;
-  for (const loc of ['it','de','en','fr']) {
-    if (scores[loc] > bestScore) { best = loc; bestScore = scores[loc]; }
-  }
-  // If all zero, default to IT (primary locale of the site).
-  return best;
-}
-
-// ─── Tokenization ─────────────────────────────────────────────────
-function normalize(s) {
-  return String(s || '')
-    .toLowerCase()
-    .normalize('NFD')
-    .replace(/[\u0300-\u036f]/g, '')
-    .replace(/[^a-z0-9\s-]/g, ' ')
-    .replace(/\s+/g, ' ')
-    .trim();
-}
-
-function tokenize(q) {
-  const norm = normalize(q);
-  if (!norm) return [];
-  return norm
-    .split(/[\s-]+/)
-    .filter(t => t.length >= 2 && !STOPWORDS.has(t));
-}
-
-// ─── Region / role token separation ────────────────────────────────
-// "Region" is any geographic term we recognize; everything else is a role token.
-const REGION_TOKENS = new Set([
-  // Ticino towns + cantons
-  'ticino','tessin','lugano','mendrisio','chiasso','bellinzona','locarno','biasca',
-  'stabio','balerna','giubiasco','massagno','manno','paradiso','melide',
-  // Swiss cantons + top cities
-  'svizzera','schweiz','swiss','suisse','switzerland',
-  'zurigo','zurich','zürich','basilea','basel','ginevra','geneve','geneva',
-  'berna','bern','lucerna','luzern','lucerne','losanna','lausanne','friburgo',
-  'friborgo','fribourg','sangallo','gallen','vallese','valais','valle','grigioni',
-  'graubunden','graubünden','neuchatel','neuenburg','wallis','jura','vaud',
-  'solothurn','soletta','uri','zug','aargau','argovia',
-  // Italian near-border (cross-border relevance)
-  'como','varese','milano','lecco','sondrio',
-  // Generic
-  'ci','ch','italia','italy','italien',
-]);
-
-function splitRoleRegion(tokens) {
-  const role = [];
-  const region = [];
-  for (const t of tokens) {
-    if (REGION_TOKENS.has(t)) region.push(t);
-    else role.push(t);
-  }
-  return { role, region };
-}
-
-function canonicalizeRegion(tokens) {
-  // Collapse synonyms so clusters merge across locales.
-  const canon = tokens.map(t => {
-    if (['ticino','tessin'].includes(t)) return 'ticino';
-    if (['svizzera','schweiz','swiss','suisse','switzerland'].includes(t)) return 'svizzera';
-    if (['zurigo','zurich','zürich'].includes(t)) return 'zurigo';
-    if (['basilea','basel'].includes(t)) return 'basilea';
-    if (['ginevra','geneve','geneva'].includes(t)) return 'ginevra';
-    if (['berna','bern'].includes(t)) return 'berna';
-    if (['lucerna','luzern','lucerne'].includes(t)) return 'lucerna';
-    if (['losanna','lausanne'].includes(t)) return 'losanna';
-    if (['grigioni','graubunden','graubünden'].includes(t)) return 'grigioni';
-    if (['vallese','valais','wallis'].includes(t)) return 'vallese';
-    if (['friburgo','friborgo','fribourg'].includes(t)) return 'friburgo';
-    return t;
-  });
-  // Dedup and sort for stable signature
-  return [...new Set(canon)].sort();
-}
-
-function canonicalizeRole(tokens) {
-  const stems = tokens.map(stemToken).filter(Boolean);
-  return [...new Set(stems)].sort();
-}
-
-// ─── Slugify for cluster canonical slug ────────────────────────────
-function slugify(s) {
-  return String(s || '')
-    .toLowerCase()
-    .normalize('NFD')
-    .replace(/[\u0300-\u036f]/g, '')
-    .replace(/[^a-z0-9]+/g, '-')
-    .replace(/^-+|-+$/g, '')
-    .slice(0, 80);
-}
+// Tokenization / stemming / locale-detection / role-region canonicalization
+// (normalize, tokenize, detectLocale, splitRoleRegion, canonicalizeRegion,
+// canonicalizeRole, slugify) live in scripts/lib/query-tokenizer.mjs — shared
+// with scripts/mine-internal-search-terms.mjs (issue #4301) so the two
+// scripts never drift on STOPWORDS/STEM_RULES/REGION_TOKENS (AGENTS.md #6).
 
 // ─── Main ─────────────────────────────────────────────────────────
 function readJsonSafe(p) {
